certificate column remove_column code refactored

+++ db/migrate/20220714092209_change_data_type_for_certificate.rb
@@ -1,3 +1,15 @@
 class ChangeDataTypeForCertificate < ActiveRecord::Migration[7.0]
-  remove_column(:doctors, :certificate)
+  def change
+    def up
+      if column_exists?(:doctors, :certificate)
+        remove_column(:doctors, :certificate)
+       end
+     end
+     
+     def down
+      if !column_exists?(:doctors, :certificate)
+       add_column :doctors, :certificate, :string
+       end
+     end
+  end
 end
